Abstract DropTarget
